backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Kanak Tanwar",email = "kanaktanwarpro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "authlib (>=1.6.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "orjson (>=3.10.18,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
